\name{muma-package}
\alias{muma-package}
\alias{muma}
\docType{package}
\title{
Metabolomics Univariate and Multivariate Analyses
}
\description{
Preprocessing of metabolomic data (normalization and scalings); Principal Component Analysis with help tool for choosing best-separating principal components and automatic outlier testing; automatic univariate analysis for parametric and non-parametric data, with report generation (volcano and box plots); partial least sqaure discriminant analysis (PLS-DA); orthogonal partial least square discriminant analysis (OPLS-DA); Statistical Total Correlation Spectroscopy (STOCSY); Ratio Analysis NMR SpectroscopY (RANSY).
}
\details{
\tabular{ll}{
Package: \tab muma\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-07-20\cr
License: \tab GPL-2\cr
}
Every analysis should start with the function 'explore.data' in order to create the proper files for subsequent analysis. However, it is possible to start the analysis also with the function 'univariate', but, in this case, only univariate analysis can be performed.
}
\author{
Edoardo Gaude, Dimitrios Spiliotopoulos, Francesca Chignola, Silvia Mari, Andrea Spitaleri and Michela Ghitti

Maintainer: Edoardo Gaude <egaude541@gmail.com>
}